\name{coef.rqpd}
\alias{coef.rqpd}
\title{Retrieve coefficients from rqpd objects}
\description{
  Alternative to using the \$ operator to retrieve coefficients from
  panel data quantile regression fits of class \code{rqpd.object}. 
}
\usage{
\method{coef}{rqpd}(object, ...)
}
\arguments{
  \item{object}{An rqpd.object.}
  \item{...}{Not used.}
}
