\name{print.rqpd.cre}
\alias{print.rqpd.cre}
\title{Print an rqpd.cre object}
\description{
Print an object generated by rqpdi with \code{panel(method="cre")}.
}
\usage{
\method{print}{rqpd.cre}(x, ...)
}
\arguments{
\item{x}{
  Object returned from rqpd representing the fit of the CRE panel data model.
 }
\item{...}{
  Optional arguments. 
 }
}
\seealso{\code{\link{rqpd}}}

\keyword{regression}
