\name{print.rqpd}
\alias{print.rqpd}
\title{Print an rqpd object}
\description{
Print an object generated by rqpd.
}
\usage{
\method{print}{rqpd}(x, ...)
}
\arguments{
\item{x}{
  Object returned from rqpd representing the fit of the panel data model.
 }
\item{...}{
  Optional arguments. 
 }
}
\seealso{\code{\link{rqpd}}}

\keyword{regression}
