\name{print.summary.rqpd}
\alias{print.summary.rqpd}
\title{Print Panel Data Quantile Regression Summary Object }
\usage{
\method{print}{summary.rqpd}(x, digits=max(5, .Options$digits - 2), ...)
}
\arguments{
  \item{x}{
    An object of class \code{"summary.rqpd"} produced by a call to 
    \code{summary.rqpd()}.
  }
  \item{digits}{
    Significant digits reported in the printed table.
  }
  \item{...}{
    Optional arguments passed to printing function
  }
}
\description{Print summary of panel data quantile regression object}
\seealso{  \code{\link{summary.rqpd}} }

\keyword{ regression }
